옵션 시그널 반복 생성 버그 수정
- _normalize_diffs: int() → float() 변환으로 "0.0" 형식 처리
- detect_action_signals: 옵션 시그널은 삭제하지 않고 값 변경 시만 갱신
- 기존 all-zero option_diffs 7,447건 NULL 변환
- 불필요 OPTION_ADD/OPTION_PRICE_CHANGE 시그널 8,259건 삭제

+++ app/master/__init__.py
@@ -15,7 +15,7 @@ def _normalize_diffs(diffs: str | None) -> str | None:
     if not diffs or not diffs.strip():
         return None
     try:
-        if all(int(v.strip()) == 0 for v in diffs.split("\n") if v.strip()):
+        if all(float(v.strip()) == 0 for v in diffs.split("\n") if v.strip()):
             return None
     except ValueError:
         pass
